factor feature_weights out of process_mth5
